requirements.txt
groq==0.9.0
gradio==4.44.0
sentence-transformers==3.1.1
faiss-cpu==1.8.0
PyMuPDF==1.24.10
networkx==3.3
matplotlib==3.9.2
Pillow==10.4.0
numpy==1.26.4
